update package.json and user component
